NewGoodsFragment数据加载

+++ src/cn/ucai/fuliCenter/bean/NewGoodBean.java
@@ -49,7 +49,7 @@ public class NewGoodBean implements Serializable {
     private String colorName;
     private String colorCode;
     private String colorUrl;
-    private int addTime;
+    private Long addTime;
 
     public int getId() {
         return id;
@@ -187,11 +187,11 @@ public class NewGoodBean implements Serializable {
         this.colorUrl = colorUrl;
     }
 
-    public int getAddTime() {
+    public Long getAddTime() {
         return addTime;
     }
 
-    public void setAddTime(int addTime) {
+    public void setAddTime(Long addTime) {
         this.addTime = addTime;
     }
